Functions that are to be moved to //shared/htrutils captured in separate file.
